Add the ability to create files

+++ controllers/utilities/removeAllUploads.ts
@@ -5,6 +5,9 @@ export default async function removeAllUploads() {
     const uploadDirectory = 'uploads';
 
     for (const file of await fs.readdir(uploadDirectory)) {
+        if (file === 'placeholder') {
+            continue;
+        }
         await fs.unlink(path.join(uploadDirectory, file));
     }
 }
